consolidate preferences into common
